move MyTuple to its own header and source file
