+++ Training/src/TMVAClassificationAppNtuple.cc
@@ -56,6 +56,7 @@
 
 #include "Ana/Common.h"
 #include "Ana/TreeHelpers.h"
+#include "Ana/MyNTuple.h"
 
 using std::map;
 using std::tuple;
