+++ Utilities/Linkdef.h
@@ -11,6 +11,8 @@
 #pragma link C++ class ParticleTreeMC+;
 #pragma link C++ class ParticleTreeMC2+;
 
+#pragma link C++ class MyNTuple+;
+
 #pragma link C++ class MatchCriterion+;
 
 #pragma link C++ class EfficiencyTable<TGraph>+;
